tests: moved utils out of conftest.py
