+++ tests/integration/conftest.py
@@ -8,39 +8,34 @@ from _pytest.fixtures import SubRequest
 from aiohttp import ClientSession
 from aiohttp.pytest_plugin import AiohttpClient
 from contextlib import closing, ExitStack
-from dataclasses import asdict, dataclass
+from dataclasses import asdict
 from filelock import FileLock
 from kafka import KafkaProducer
-from kafka.errors import LeaderNotAvailableError, NoBrokersAvailable
 from karapace.client import Client
 from karapace.config import set_config_defaults, write_config
 from karapace.kafka_rest_apis import KafkaRest, KafkaRestAdminClient
 from karapace.schema_registry_apis import KarapaceSchemaRegistry
-from karapace.utils import Expiration
 from pathlib import Path
 from subprocess import Popen
+from tests.integration.utils.config import KafkaDescription, ZKConfig
+from tests.integration.utils.kafka_server import configure_and_start_kafka, maybe_download_kafka, wait_for_kafka
+from tests.integration.utils.process import stop_process, wait_for_port_subprocess
+from tests.integration.utils.synchronization import lock_path_for
+from tests.integration.utils.zookeeper import configure_and_start_zk
 from tests.utils import (
     get_random_port,
-    KAFKA_PORT_RANGE,
     KafkaConfig,
     KafkaServers,
     new_random_name,
     REGISTRY_PORT_RANGE,
     repeat_until_successful_request,
-    ZK_PORT_RANGE,
 )
-from typing import AsyncIterator, Dict, Iterator, List, Optional, Tuple
+from typing import AsyncIterator, Iterator, Optional, Tuple
 
 import asyncio
-import logging
 import os
 import pathlib
 import pytest
-import requests
-import signal
-import socket
-import tarfile
-import time
 import ujson
 
 REPOSITORY_DIR = pathlib.Path(__file__).parent.parent.parent.absolute()
@@ -48,122 +43,6 @@ RUNTIME_DIR = REPOSITORY_DIR / "runtime"
 TEST_INTEGRATION_DIR = REPOSITORY_DIR / "tests" / "integration"
 KAFKA_WAIT_TIMEOUT = 60
 KAFKA_SCALA_VERSION = "2.13"
-
-
-@dataclass
-class ZKConfig:
-    client_port: int
-    admin_port: int
-    path: str
-
-    @staticmethod
-    def from_dict(data: dict) -> "ZKConfig":
-        return ZKConfig(
-            data["client_port"],
-            data["admin_port"],
-            data["path"],
-        )
-
-
-@dataclass(frozen=True)
-class KafkaDescription:
-    version: str
-    install_dir: Path
-    download_url: str
-    protocol_version: str
-
-
-def stop_process(proc: Optional[Popen]) -> None:
-    if proc:
-        os.kill(proc.pid, signal.SIGKILL)
-        proc.wait(timeout=10.0)
-
-
-def port_is_listening(hostname: str, port: int, ipv6: bool) -> bool:
-    if ipv6:
-        s = socket.socket(socket.AF_INET6, socket.SOCK_STREAM, 0)
-    else:
-        s = socket.socket()
-    s.settimeout(0.5)
-    try:
-        s.connect((hostname, port))
-        s.close()
-        return True
-    except socket.error:
-        return False
-
-
-def wait_for_kafka(
-    kafka_servers: KafkaServers,
-    wait_time: float,
-) -> None:
-    for server in kafka_servers.bootstrap_servers:
-        expiration = Expiration.from_timeout(timeout=wait_time)
-
-        list_topics_successful = False
-        while not list_topics_successful:
-            expiration.raise_timeout_if_expired(
-                msg_format="Could not contact kafka cluster on host `{server}`",
-                server=server,
-            )
-            try:
-                KafkaRestAdminClient(bootstrap_servers=server).cluster_metadata()
-            # ValueError:
-            # - if the port number is invalid (i.e. not a number)
-            # - if the port is not bound yet
-            # NoBrokersAvailable:
-            # - if the address/port does not point to a running server
-            # LeaderNotAvailableError:
-            # - if there is no leader yet
-            except (
-                NoBrokersAvailable,
-                LeaderNotAvailableError,
-                ValueError,
-            ) as e:
-                print(f"Error checking kafka cluster: {e}")
-                time.sleep(2.0)
-            else:
-                list_topics_successful = True
-
-
-def wait_for_port_subprocess(
-    port: int,
-    process: Popen,
-    *,
-    hostname: str = "127.0.0.1",
-    wait_time: float = 20.0,
-    ipv6: bool = False,
-) -> None:
-    expiration = Expiration.from_timeout(wait_time)
-
-    while not port_is_listening(hostname, port, ipv6):
-        expiration.raise_timeout_if_expired(
-            msg_format="Timeout waiting for `{hostname}:{port}`",
-            hostname=hostname,
-            port=port,
-        )
-        assert process.poll() is None, f"Process no longer running, exit_code: {process.returncode}"
-        time.sleep(2.0)
-
-    elapsed = expiration.elapsed
-    print(f"Server `{hostname}:{port}` listening after {elapsed} seconds")
-
-
-def lock_path_for(path: Path) -> Path:
-    """Append .lock to path"""
-    suffixes = path.suffixes
-    suffixes.append(".lock")
-    return path.with_suffix("".join(suffixes))
-
-
-def maybe_download_kafka(kafka_description: KafkaDescription) -> None:
-    """If necessary download kafka to run the tests."""
-    if not os.path.exists(kafka_description.install_dir):
-        logging.info("Downloading Kafka {url}", url=kafka_description.download_url)
-
-        download = requests.get(kafka_description.download_url, stream=True)
-        with tarfile.open(mode="r:gz", fileobj=download.raw) as file:
-            file.extractall(str(kafka_description.install_dir.parent))
 
 
 @pytest.fixture(scope="session", name="kafka_description")
@@ -222,11 +101,13 @@ def fixture_kafka_server(
                 # Make sure zookeeper is running before trying to start Kafka
                 wait_for_port_subprocess(zk_config.client_port, zk_proc, wait_time=20)
 
+                log4j_config = TEST_INTEGRATION_DIR / "config" / "log4j.properties"
                 kafka_config, kafka_proc = configure_and_start_kafka(
-                    session_datadir,
-                    session_logdir,
-                    zk_config,
-                    kafka_description,
+                    datadir=session_datadir,
+                    logdir=session_logdir,
+                    log4j_config=log4j_config,
+                    zk=zk_config,
+                    kafka_description=kafka_description,
                 )
                 stack.callback(stop_process, kafka_proc)
 
@@ -547,211 +428,3 @@ async def fixture_registry_async_client_tls(
         yield client
     finally:
         await client.close()
-
-
-def zk_java_args(cfg_path: Path, kafka_description: KafkaDescription) -> List[str]:
-    msg = f"Couldn't find kafka installation at {kafka_description.install_dir} to run integration tests."
-    assert kafka_description.install_dir.exists(), msg
-    java_args = [
-        "-cp",
-        str(kafka_description.install_dir / "libs" / "*"),
-        "org.apache.zookeeper.server.quorum.QuorumPeerMain",
-        str(cfg_path),
-    ]
-    return java_args
-
-
-def kafka_java_args(
-    heap_mb: int,
-    kafka_config_path: str,
-    logs_dir: str,
-    log4j_properties_path: str,
-    kafka_description: KafkaDescription,
-) -> List[str]:
-    msg = f"Couldn't find kafka installation at {kafka_description.install_dir} to run integration tests."
-    assert kafka_description.install_dir.exists(), msg
-    java_args = [
-        "-Xmx{}M".format(heap_mb),
-        "-Xms{}M".format(heap_mb),
-        "-Dkafka.logs.dir={}/logs".format(logs_dir),
-        "-Dlog4j.configuration=file:{}".format(log4j_properties_path),
-        "-cp",
-        str(kafka_description.install_dir / "libs" / "*"),
-        "kafka.Kafka",
-        kafka_config_path,
-    ]
-    return java_args
-
-
-def get_java_process_configuration(java_args: List[str]) -> List[str]:
-    command = [
-        "/usr/bin/java",
-        "-server",
-        "-XX:+UseG1GC",
-        "-XX:MaxGCPauseMillis=20",
-        "-XX:InitiatingHeapOccupancyPercent=35",
-        "-XX:+DisableExplicitGC",
-        "-XX:+ExitOnOutOfMemoryError",
-        "-Djava.awt.headless=true",
-        "-Dcom.sun.management.jmxremote",
-        "-Dcom.sun.management.jmxremote.authenticate=false",
-        "-Dcom.sun.management.jmxremote.ssl=false",
-    ]
-    command.extend(java_args)
-    return command
-
-
-def configure_and_start_kafka(
-    datadir: Path,
-    logdir: Path,
-    zk: ZKConfig,
-    kafka_description: KafkaDescription,
-) -> Tuple[KafkaConfig, Popen]:
-    # setup filesystem
-    data_dir = datadir / "kafka"
-    log_dir = logdir / "kafka"
-    config_path = log_dir / "server.properties"
-    data_dir.mkdir(parents=True)
-    log_dir.mkdir(parents=True)
-
-    plaintext_port = get_random_port(port_range=KAFKA_PORT_RANGE, blacklist=[])
-
-    config = KafkaConfig(
-        datadir=str(data_dir),
-        kafka_keystore_password="secret",
-        kafka_port=plaintext_port,
-        zookeeper_port=zk.client_port,
-    )
-
-    advertised_listeners = ",".join(
-        [
-            "PLAINTEXT://127.0.0.1:{}".format(plaintext_port),
-        ]
-    )
-    listeners = ",".join(
-        [
-            "PLAINTEXT://:{}".format(plaintext_port),
-        ]
-    )
-
-    # Keep in sync with containers/docker-compose.yml
-    kafka_config = {
-        "broker.id": 1,
-        "broker.rack": "local",
-        "advertised.listeners": advertised_listeners,
-        "auto.create.topics.enable": False,
-        "default.replication.factor": 1,
-        "delete.topic.enable": "true",
-        "inter.broker.listener.name": "PLAINTEXT",
-        "inter.broker.protocol.version": kafka_description.protocol_version,
-        "listeners": listeners,
-        "log.cleaner.enable": "true",
-        "log.dirs": config.datadir,
-        "log.message.format.version": kafka_description.protocol_version,
-        "log.retention.check.interval.ms": 300000,
-        "log.segment.bytes": 200 * 1024 * 1024,  # 200 MiB
-        "log.preallocate": False,
-        "num.io.threads": 8,
-        "num.network.threads": 112,
-        "num.partitions": 1,
-        "num.replica.fetchers": 4,
-        "num.recovery.threads.per.data.dir": 1,
-        "offsets.topic.replication.factor": 1,
-        "socket.receive.buffer.bytes": 100 * 1024,
-        "socket.request.max.bytes": 100 * 1024 * 1024,
-        "socket.send.buffer.bytes": 100 * 1024,
-        "transaction.state.log.min.isr": 1,
-        "transaction.state.log.num.partitions": 16,
-        "transaction.state.log.replication.factor": 1,
-        "zookeeper.connection.timeout.ms": 6000,
-        "zookeeper.connect": f"127.0.0.1:{zk.client_port}",
-    }
-
-    with config_path.open("w") as fp:
-        for key, value in kafka_config.items():
-            fp.write("{}={}\n".format(key, value))
-
-    # stdout logger is disabled to keep the pytest report readable
-    log4j_properties_path = str(TEST_INTEGRATION_DIR / "config" / "log4j.properties")
-
-    kafka_cmd = get_java_process_configuration(
-        java_args=kafka_java_args(
-            heap_mb=256,
-            logs_dir=str(log_dir),
-            log4j_properties_path=log4j_properties_path,
-            kafka_config_path=str(config_path),
-            kafka_description=kafka_description,
-        ),
-    )
-    env: Dict[bytes, bytes] = {}
-    proc = Popen(kafka_cmd, env=env)
-    return config, proc
-
-
-def configure_and_start_zk(zk_dir: Path, kafka_description: KafkaDescription) -> Tuple[ZKConfig, Popen]:
-    cfg_path = zk_dir / "zoo.cfg"
-    logs_dir = zk_dir / "logs"
-    logs_dir.mkdir(parents=True)
-
-    client_port = get_random_port(port_range=ZK_PORT_RANGE, blacklist=[])
-    admin_port = get_random_port(port_range=ZK_PORT_RANGE, blacklist=[client_port])
-    config = ZKConfig(
-        client_port=client_port,
-        admin_port=admin_port,
-        path=str(zk_dir),
-    )
-    zoo_cfg = """
-# The number of milliseconds of each tick
-tickTime=2000
-# The number of ticks that the initial
-# synchronization phase can take
-initLimit=10
-# The number of ticks that can pass between
-# sending a request and getting an acknowledgement
-syncLimit=5
-# the directory where the snapshot is stored.
-# do not use /tmp for storage, /tmp here is just
-# example sakes.
-dataDir={path}
-# the port at which the clients will connect
-clientPort={client_port}
-#clientPortAddress=127.0.0.1
-# the maximum number of client connections.
-# increase this if you need to handle more clients
-#maxClientCnxns=60
-#
-# Be sure to read the maintenance section of the
-# administrator guide before turning on autopurge.
-#
-# http://zookeeper.apache.org/doc/current/zookeeperAdmin.html#sc_maintenance
-#
-# The number of snapshots to retain in dataDir
-#autopurge.snapRetainCount=3
-# Purge task interval in hours
-# Set to "0" to disable auto purge feature
-#autopurge.purgeInterval=1
-# admin server
-admin.serverPort={admin_port}
-admin.enableServer=false
-# Allow reconfig calls to be made to add/remove nodes to the cluster on the fly
-reconfigEnabled=true
-# Don't require authentication for reconfig
-skipACL=yes
-""".format(
-        client_port=config.client_port,
-        admin_port=config.admin_port,
-        path=config.path,
-    )
-    cfg_path.write_text(zoo_cfg)
-    env = {
-        "CLASSPATH": "/usr/share/java/slf4j/slf4j-simple.jar",
-        "ZOO_LOG_DIR": str(logs_dir),
-    }
-    java_args = get_java_process_configuration(
-        java_args=zk_java_args(
-            cfg_path,
-            kafka_description,
-        )
-    )
-    proc = Popen(java_args, env=env)
-    return config, proc
