+++ tests/integration/utils/network.py
@@ -1,0 +1,19 @@
+"""
+Copyright (c) 2022 Aiven Ltd
+See LICENSE for details
+"""
+import socket
+
+
+def port_is_listening(hostname: str, port: int, ipv6: bool) -> bool:
+    if ipv6:
+        s = socket.socket(socket.AF_INET6, socket.SOCK_STREAM, 0)
+    else:
+        s = socket.socket()
+    s.settimeout(0.5)
+    try:
+        s.connect((hostname, port))
+        s.close()
+        return True
+    except socket.error:
+        return False
