+++ tests/integration/utils/synchronization.py
@@ -1,0 +1,12 @@
+"""
+Copyright (c) 2022 Aiven Ltd
+See LICENSE for details
+"""
+from pathlib import Path
+
+
+def lock_path_for(path: Path) -> Path:
+    """Append .lock to path"""
+    suffixes = path.suffixes
+    suffixes.append(".lock")
+    return path.with_suffix("".join(suffixes))
